requirements.txt
torch
opencv-python
dronekit
numpy
pandas
folium
python-telegram-bot
PyDrive
